Refactor bucket ring into its own package
