+++ goldmane/pkg/emitter/emitter.go
@@ -31,6 +31,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/projectcalico/calico/goldmane/pkg/aggregator"
+	"github.com/projectcalico/calico/goldmane/pkg/aggregator/bucketing"
 )
 
 var (
@@ -128,7 +129,7 @@ func (e *Emitter) Run(stopCh chan struct{}) {
 	}
 }
 
-func (e *Emitter) Receive(bucket *aggregator.FlowCollection) {
+func (e *Emitter) Receive(bucket *bucketing.FlowCollection) {
 	// Add the bucket to our internal map so we can retry it if needed.
 	// We'll remove it from the map once it's successfully emitted.
 	k := bucketKey{startTime: bucket.StartTime, endTime: bucket.EndTime}
@@ -154,7 +155,7 @@ func (e *Emitter) forget(k bucketKey) {
 	e.q.Forget(k)
 }
 
-func (e *Emitter) emit(bucket *aggregator.FlowCollection) error {
+func (e *Emitter) emit(bucket *bucketing.FlowCollection) error {
 	// Check if we have already emitted this batch. If it pre-dates
 	// the latest timestamp we've emitted, skip it. This can happen, for example, on restart when
 	// we learn already emitted flows from the cache.
@@ -164,7 +165,7 @@ func (e *Emitter) emit(bucket *aggregator.FlowCollection) error {
 	}
 
 	// Marshal the flows to JSON and send them to the emitter.
-	rdr, err := e.bucketToReader(bucket)
+	rdr, err := e.collectionToReader(bucket)
 	if err != nil {
 		return err
 	}
@@ -182,7 +183,7 @@ func (e *Emitter) emit(bucket *aggregator.FlowCollection) error {
 	return nil
 }
 
-func (e *Emitter) bucketToReader(bucket *aggregator.FlowCollection) (*bytes.Reader, error) {
+func (e *Emitter) collectionToReader(bucket *bucketing.FlowCollection) (*bytes.Reader, error) {
 	body := []byte{}
 	for _, flow := range bucket.Flows {
 		if len(body) != 0 {
